Rework dependencies so that contacts package import indexer
